requirements/common.txt
SQLAlchemy==1.3.15
pytz
pyOpenSSL
ndg-httpsclient
pyasn1
apache-airflow[crypto,celery,postgres,hive,jdbc,mysql,ssh]==1.10.9
redis==3.2
pandas